Fix display of ellipsis in truncated text containing colour codes
`text_out_colours_ellipsis()` had a number of bugs breaking the use of ellipses for truncated text when embedded colour codes were used.

This rewrites part of `text_out_colours_ellipsis()` to fix these bugs (and make the logic easier to follow at the same time).

(More could still be done to improve the function, but that is out of scope for this change.)

(Note that some of the lines changed contain only white-space changes.)