Adjust list view scroll timer speed
reupen/columns_ui#306

This adjusts the speed of the list view scroll timer used when selecting items or during drag and drag operations so that it's faster when more items fit in the viewable area and slower when fewer fit.

It also attempts to disable automatic scrolling that was being performed by `IDropTargetHelper` which was causing the scrolling speed to be faster than intended.