Fix erasing behaviour of paint_subclassed_window_with_buffering()
This makes `paint_subclassed_window_with_buffering()` always erase the background, and do so using `WM_ERASEBKGND` instead of  `WM_PRINTCLIENT`.

This is because the memory bitmap will always start off empty (and hence erasing is always required), and as support for the `PRF_ERASEBKGND` flag of `WM_PRINTCLIENT` is very patchy.