Allow list view inline edit box to be dark themed
This makes it possible for the inline editing edit box in the list view to have a dark appearance when dark mode is enabled.

It also fixes a glitch with its right border at high DPIs and makes the initial rendering of the edit box less glitchy in some cases.