Add list view search contexts
This adds the concept of search contexts to the list view. These provide the text of items when typing directly in the list view to jump to matching items.

A custom search context can be used in cases where the default context has poor performance (i.e. virtualised lists where getting the text for an individual item is expensive).