Fix inability to click on inline editing autocomplete suggestion
reupen/columns_ui#885

This resolves a problem in the list view where it wasn't possible to click on an autocomplete suggestion when using inline editing.

This was happening because inline editing is ended when clicking outside the edit box. An exception has been added when clicking in the autocomplete drop-down so that those clicks don't end inline editing.

(Unfortunately there wasn't a particularly clean way to implement this, so this will have to do.)

This also resolves a problem where inline editing wasn't closed when clicking on non-client areas of the UI (such as a scroll bar or the title bar).