Add additional font family and face look-up helpers
This makes some changes to the DirectWrite helpers to facilitate serialisation and deserialisation of font family names and weight, stretch and style properties (and hence recreating text formats after deserialisation).

Furthermore, adjustments were made to the list view so that DirectWrite text formats can be specified where DirectWrite is used for text rendering.