Render list view header text using DirectWrite
This uses a few tricks to get the column titles in the list view header rendered using DirectWrite.

No text is passed to the header control for each column, to stop the control rendering the text. The text can then be rendered using custom draw using DirectWrite.

Naturally this is a hack. (It may have a negative effect on accessibility, however accessibility is a general problem for the list view control already.)