Fix occasional DirectWrite stack overflow
This resolves the DirectWrite crash mentioned in https://blog.yuo.be/2024/12/07/the-unresolved-crash-releasing-an-idwritetextformat/.

This was traced to calling `IDWriteFactory::CreateEllipsisTrimmingSign()` with a text format, and then passing the returned `IDWriteInlineObject` to `IDWriteTextFormat::SetTrimming()` for the same text format. Evidently, this was causing some recursion when releasing the text format, which sometimes caused a stack overflow.

The fix is to simply call `IDWriteTextFormat::SetTrimming()` on the `IDWriteTextLayout` object  instead (`IDWriteTextLayout` inherits from `IDWriteTextFormat`). From testing, this breaks the recursion. Doing this on the text layout is also more logical and convenient.