Work around DirectWrite not rendering trailing whitespace
This adds a somewhat crude workaround for DirectWrite not rendering trailing whitespace for centre- and right-aligned text.

The workaround is to add a zero-width space at the end of the string. (A zero-width space is not considered a whitespace character.)

Unfortunately, I've found no other solution for this problem (other than using custom text layout code, which is not a particularly enticing prospect).

The workaround is only useful for single-line text, and is part of the `text_out_columns_and_colours()` helper (rather than the underlying DirectWrite helpers, as it doesn't make sense to do this there).