Fix DirectWrite colour effects not being applied to ellipses
This works around a problem with the inline object created by `IDWriteFactory::CreateEllipsisTrimmingSign()` seemingly does nothing with the `clientDrawingEffect` passed to it, causing the ellipsis to be rendered using the wrong colour when a colour drawing effect was active.

(This change is only relevant for the custom GDI renderer. Interestingly the renderer used by Direct2D does not have this problem.)