Optimise list view navigation using arrow keys
This reduces CPU usage when navigating in the list view using the Up and Down keys.

(Previously, it suspended redrawing while handling the key press, to avoid visual artefacts. However, that meant the entire client area needed redrawing afterwards, which was unnecessary work. This removes the drawing suspension and tweaks the logic to avoid any visual artefacts.)