Add MenuCommandCollector class
This adds a `MenuCommandCollector` class for collecting menu item command callback functions and generating menu item IDs for them, to avoid manual menu item ID assignment and other logic which is all error prone.

This is separated from the `Menu` class, so that it can be used across submenus.

`Menu::run()` has also been changed to return a `uint32_t`.