Use custom drag images on Windows 10 and newer
This updates the list view to use custom drag images on Windows 10 and newer.

This is to avoid the default white drag image introduced on Windows 11, and also to have a dark themed drag image.

There is no change on older versions of Windows.

Some Direct2D and Direct3D utilities have been moved from Columns UI to here.