Add the ability for the list view group info area to be sticky
This adds the ability for the list view group info area to be sticky, so that it stays in view when scrolling down while the group is still on screen.

Some rendering code was updated to accomodate this without much of a performance hit. Rendering should generally be slightly more efficient as a result.

A bug where the drag-and-drop insertion marker was rendered in the wrong place when the list view was scrolled horizontally was also fixed.