Add LRU cache for DirectWrite text layouts
This adds an LRU cache for DirectWrite text layouts to improve performance when the same text is rendered multiple times.

Note that the text layout max width and height is part of the cache key, as it seems that DirectWrite flushes some cached metrics when changing the max dimensions (even if it wouldn't cause a layout change).