Add ability to disable list view group info area vertical spacing
This adds the ability to disable additional list view spacing below leaf group headers and after the last item of a group when the group info area is enabled.

A problem where, when the extra spacing is enabled, it was not present for the first group was also fixed.

A minor error in hit-testing calculations was also fixed.