Improve consistency of list view group visibility logic
This makes some changes to make the logic around whether a group is shown or hidden more consistent. In particular, the short-circuiting in `ListView::get_item_display_group_count()` was removed as the rendering logic doesn’t short-circuit in the same way.

In theory none of this should make a difference, as if child groups different between a parent group then the parent groups should too. However, a crash was submitted indicating that there is some scenario where it does, indicating there may be a bug somewhere else. So far, I haven’t been able to find a scenario where things end up in a state where it matters, though.

Additionally a bug when replacing items causing display indices not to be set correctly for a hidden group was fixed, and the `ListView::replace_items()` method now returns whether display indices for subsequent items changed due to visible group header counts changing.